+++ gaphor/SysML/diagramitems.py
@@ -1,2 +1,3 @@
 from gaphor.SysML.blocks import *
 from gaphor.SysML.requirements import *
+from gaphor.SysML.diagramframe import DiagramFrameItem